+++ ui/packages/shared/profile/src/useQuery.tsx
@@ -26,6 +26,7 @@ export interface IQueryResult {
 
 interface UseQueryOptions {
   skip?: boolean;
+  nodeTrimThreshold?: number;
 }
 
 export const useQuery = (
@@ -37,10 +38,11 @@ export const useQuery = (
   const {skip = false} = options ?? {};
   const metadata = useGrpcMetadata();
   const {data, isLoading, error} = useGrpcQuery<QueryResponse | undefined>({
-    key: ['query', profileSource, reportType],
+    key: ['query', profileSource, reportType, options?.nodeTrimThreshold],
     queryFn: async () => {
       const req = profileSource.QueryRequest();
       req.reportType = reportType;
+      req.nodeTrimThreshold = options?.nodeTrimThreshold;
 
       const {response} = await client.query(req, {meta: metadata});
       return response;