+++ ui/packages/shared/client/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.62 (2022-12-22)
+
+**Note:** Version bump only for package @parca/client
+
 ## 0.16.61 (2022-12-20)
 
 # 0.15.0 (2022-12-20)